feat: make the logger configurable via environment variables.
#19
We decided to use the simiple approach that is provided by spdlog for now. We need to implement reading config files anyway at a later point. We might think about putting the logger config in those config files, too.
