+++ Dockerfile
@@ -35,4 +35,5 @@ COPY --from=builder /src/siloApi .
 # call /info, extract "seqeunceCount" from the JSON and assert that the value is not 0. If any of those fails, "exit 1".
 HEALTHCHECK CMD curl --fail --silent localhost:8080/info | jq .sequenceCount | xargs test 0 -ne || exit 1
 
+ENV SPDLOG_LEVEL="off,file_logger=debug"
 CMD ["./siloApi", "--api"]
