+++ README.md
@@ -64,14 +64,31 @@ build/silo_test
 ```
 
 For linting we use clang-tidy. The config is stored in `.clang-tidy`. It will run automatically with the build process
-and will throw errors accordingly. However it is rather slow. If you only want a fast build use
+and will throw errors accordingly. However, it is rather slow. If you only want a fast build use
 
 ```shell
 ./build_with_conan build_without_clang_tidy
 ```
 
-When pushing to github, a separate Docker image will be built, which runs the formatter. (This is a workaround, because
+When pushing to GitHub, a separate Docker image will be built, which runs the formatter. (This is a workaround, because
 building with clang-tidy under alpine was not possible yet.)
+
+# Logging
+
+We use [spdlog](https://github.com/gabime/spdlog) for logging. The log level and log target can be controlled via
+environment variables:
+
+* Start SILO with `SPDLOG_LEVEL=off,console_logger=debug` to log to stdout at debug level.
+* Start SILO with `SPDLOG_LEVEL=off,file_logger=trace` to log to daily rotating files in `log/` (relative from the
+  location where SILO is executed) at trace level.
+
+If `SPDLOG_LEVEL` is not set, SILO will log to the console at info level.
+
+We decided to use the macros provided by spdlog rather than the functions, because this lets us disable log statements
+at compile time by adjusting `add_compile_definitions(SPDLOG_ACTIVE_LEVEL=SPDLOG_LEVEL_TRACE)` to the desired log level
+via CMake. This might be desirable for benchmarking SILO. However, the default should be `SPDLOG_LEVEL_TRACE` to give
+the maintainer the possibility to adjust the log level to a log level that they prefer, without the need to recompile
+SILO.
 
 # Code Style Guidelines
