+++ include/silo_api/logging.h
@@ -1,0 +1,6 @@
+#ifndef SILO_LOGGING_H
+#define SILO_LOGGING_H
+
+void setupLogger();
+
+#endif  // SILO_LOGGING_H
